requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.6.1
Automat==20.2.0
Brotli==1.0.7
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
Django==3.0.7
django-autoslug-iplweb==1.9.5
django-channels==0.7.0
django-crispy-forms==1.9.1
django-multiselectfield==0.1.12
django-nine==0.2.3
django-plotly-dash==1.3.1
django-tinymce==3.0.2
dpd-components==0.1.0
dpd-static-support==0.0.5
easy-thumbnails==2.7
feedparser==5.2.1
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msgpack==0.6.2
numpy==1.18.5
oauthlib==3.1.0
ordereddict==1.1
pandas==1.0.5
pif==0.8.2
Pillow==7.1.2
plotly==4.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
retrying==1.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
transliterate==1.10.2
Twisted==20.3.0
txaio==20.4.1
Unidecode==1.1.1
urllib3==1.25.9
vishap==0.1.5
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
zope.interface==5.1.0
